+++ YoutubeExplode/Bridge/PlayerSource.cs
@@ -11,86 +11,86 @@ internal partial class PlayerSource
 {
     private readonly string _content;
 
-    public string? SignatureTimestamp => Memo.Cache(this, () =>
-        Regex.Match(_content, @"(?:signatureTimestamp|sts):(\d{5})")
+    public CipherManifest? CipherManifest => Memo.Cache(this, () =>
+    {
+        // Extract the signature timestamp
+        var signatureTimestamp = Regex.Match(_content, @"(?:signatureTimestamp|sts):(\d{5})")
             .Groups[1]
             .Value
-            .NullIfWhiteSpace()
-    );
+            .NullIfWhiteSpace();
 
-    private string? CipherCallsite => Memo.Cache(this, () =>
-        Regex.Match(
+        if (string.IsNullOrWhiteSpace(signatureTimestamp))
+            return null;
+
+        // Find where the player calls the cipher functions
+        var cipherCallsite = Regex.Match(
             _content,
             """
-            \w+=function\(\w+\){(\w+)=\1\.split\(['"]{2}\);.*?return \1\.join\(['"]{2}\)}
-            """,
-            RegexOptions.Singleline
-        ).Groups[0].Value.NullIfWhiteSpace()
-    );
-
-    private string? CipherDefinition => Memo.Cache(this, () =>
-    {
-        if (string.IsNullOrWhiteSpace(CipherCallsite))
-            return null;
-
-        var objName = Regex.Match(CipherCallsite, @"(\w+)\.\w+\(\w+,\d+\);")
-            .Groups[1]
-            .Value;
-
-        if (string.IsNullOrWhiteSpace(objName))
-            return null;
-
-        return Regex.Match(
-            _content,
-            $$"""
-            var {{Regex.Escape(objName)}}={.*?};
+            [$_\w]+=function\([$_\w]+\){([$_\w]+)=\1\.split\(['"]{2}\);.*?return \1\.join\(['"]{2}\)}
             """,
             RegexOptions.Singleline
         ).Groups[0].Value.NullIfWhiteSpace();
-    });
 
-    public CipherManifest? CipherManifest => Memo.Cache(this, () =>
-    {
-        if (string.IsNullOrWhiteSpace(CipherCallsite))
+        if (string.IsNullOrWhiteSpace(cipherCallsite))
             return null;
 
-        if (string.IsNullOrWhiteSpace(CipherDefinition))
+        // Find the object that defines the cipher functions
+        var cipherContainerName = Regex.Match(cipherCallsite, @"([$_\w]+)\.[$_\w]+\([$_\w]+,\d+\);")
+            .Groups[1]
+            .Value;
+
+        if (string.IsNullOrWhiteSpace(cipherContainerName))
             return null;
 
+        // Find the definition of the cipher functions
+        var cipherDefinition = Regex.Match(
+            _content,
+            $$"""
+            var {{Regex.Escape(cipherContainerName)}}={.*?};
+            """,
+            RegexOptions.Singleline
+        ).Groups[0].Value.NullIfWhiteSpace();
+
+        if (string.IsNullOrWhiteSpace(cipherDefinition))
+            return null;
+
+        // Identify the swap cipher function
         var swapFuncName = Regex.Match(
-            CipherDefinition,
-            @"(\w+):function\(\w+,\w+\){+[^}]*?%[^}]*?}",
+            cipherDefinition,
+            @"([$_\w]+):function\([$_\w]+,[$_\w]+\){+[^}]*?%[^}]*?}",
             RegexOptions.Singleline
         ).Groups[1].Value.NullIfWhiteSpace();
 
+        // Identify the splice cipher function
         var spliceFuncName = Regex.Match(
-            CipherDefinition,
-            @"(\w+):function\(\w+,\w+\){+[^}]*?splice[^}]*?}",
+            cipherDefinition,
+            @"([$_\w]+):function\([$_\w]+,[$_\w]+\){+[^}]*?splice[^}]*?}",
             RegexOptions.Singleline
         ).Groups[1].Value.NullIfWhiteSpace();
 
+        // Identify the reverse cipher function
         var reverseFuncName = Regex.Match(
-            CipherDefinition,
-            @"(\w+):function\(\w+\){+[^}]*?reverse[^}]*?}",
+            cipherDefinition,
+            @"([$_\w]+):function\([$_\w]+\){+[^}]*?reverse[^}]*?}",
             RegexOptions.Singleline
         ).Groups[1].Value.NullIfWhiteSpace();
 
         var operations = new List<ICipherOperation>();
 
-        foreach (var statement in CipherCallsite.Split(";"))
+        foreach (var statement in cipherCallsite.Split(';'))
         {
-            var calledFuncName = Regex.Match(statement, @"\w+\.(\w+)\(\w+,\d+\)").Groups[1].Value;
+            var calledFuncName = Regex.Match(statement, @"[$_\w]+\.([$_\w]+)\([$_\w]+,\d+\)").Groups[1].Value;
             if (string.IsNullOrWhiteSpace(calledFuncName))
                 continue;
 
             if (string.Equals(calledFuncName, swapFuncName, StringComparison.Ordinal))
             {
-                var index = Regex.Match(statement, @"\(\w+,(\d+)\)").Groups[1].Value.ParseInt();
+                var index = Regex.Match(statement, @"\([$_\w]+,(\d+)\)").Groups[1].Value.ParseInt();
                 operations.Add(new SwapCipherOperation(index));
             }
             else if (string.Equals(calledFuncName, spliceFuncName, StringComparison.Ordinal))
             {
-                var index = Regex.Match(statement, @"\(\w+,(\d+)\)").Groups[1].Value.ParseInt();
+                var index = Regex.Match(statement, @"\([$_\w]+,(\d+)\)").Groups[1].Value.ParseInt();
                 operations.Add(new SpliceCipherOperation(index));
             }
             else if (string.Equals(calledFuncName, reverseFuncName, StringComparison.Ordinal))
@@ -99,7 +99,7 @@ internal partial class PlayerSource
             }
         }
 
-        return new CipherManifest(operations);
+        return new CipherManifest(signatureTimestamp, operations);
     });
 
     public PlayerSource(string content) => _content = content;