+++ YoutubeExplode/Playlists/PlaylistController.cs
@@ -3,7 +3,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using YoutubeExplode.Bridge;
 using YoutubeExplode.Exceptions;
-using YoutubeExplode.Utils;
 using YoutubeExplode.Videos;
 
 namespace YoutubeExplode.Playlists;
@@ -21,21 +20,23 @@ internal class PlaylistController
     {
         using var request = new HttpRequestMessage(HttpMethod.Post, "https://www.youtube.com/youtubei/v1/browse")
         {
-            Content = Json.SerializeToHttpContent(new
-            {
-                browseId = "VL" + playlistId,
-                context = new
+            Content = new StringContent(
+                // lang=json
+                $$"""
                 {
-                    client = new
-                    {
-                        clientName = "WEB",
-                        clientVersion = "2.20210408.08.00",
-                        hl = "en",
-                        gl = "US",
-                        utcOffsetMinutes = 0
+                    "browseId": "VL{{playlistId}}",
+                    "context": {
+                        "client": {
+                            "clientName": "WEB",
+                            "clientVersion": "2.20210408.08.00",
+                            "hl": "en",
+                            "gl": "US",
+                            "utcOffsetMinutes": 0
+                        }
                     }
                 }
-            })
+                """
+            )
         };
 
         using var response = await _http.SendAsync(request, cancellationToken);
@@ -63,24 +64,26 @@ internal class PlaylistController
         {
             using var request = new HttpRequestMessage(HttpMethod.Post, "https://www.youtube.com/youtubei/v1/next")
             {
-                Content = Json.SerializeToHttpContent(new
-                {
-                    playlistId = playlistId.Value,
-                    videoId = videoId?.Value,
-                    playlistIndex = index,
-                    context = new
+                Content = new StringContent(
+                    // lang=json
+                    $$"""
                     {
-                        client = new
-                        {
-                            clientName = "WEB",
-                            clientVersion = "2.20210408.08.00",
-                            hl = "en",
-                            gl = "US",
-                            utcOffsetMinutes = 0,
-                            visitorData
+                        "playlistId": "{{playlistId}}",
+                        "videoId": "{{videoId}}",
+                        "playlistIndex": {{index}},
+                        "context": {
+                            "client": {
+                                "clientName": "WEB",
+                                "clientVersion": "2.20210408.08.00",
+                                "hl": "en",
+                                "gl": "US",
+                                "utcOffsetMinutes": 0,
+                                "visitorData": "{{visitorData}}"
+                            }
                         }
                     }
-                })
+                    """
+                )
             };
 
             using var response = await _http.SendAsync(request, cancellationToken);