+++ YoutubeExplode/Search/SearchController.cs
@@ -2,7 +2,6 @@
 using System.Threading;
 using System.Threading.Tasks;
 using YoutubeExplode.Bridge;
-using YoutubeExplode.Utils;
 
 namespace YoutubeExplode.Search;
 
@@ -20,29 +19,31 @@ internal class SearchController
     {
         using var request = new HttpRequestMessage(HttpMethod.Post, "https://www.youtube.com/youtubei/v1/search")
         {
-            Content = Json.SerializeToHttpContent(new
-            {
-                query = searchQuery,
-                @params = searchFilter switch
+            Content = new StringContent(
+                // lang=json
+                $$"""
                 {
-                    SearchFilter.Video => "EgIQAQ%3D%3D",
-                    SearchFilter.Playlist => "EgIQAw%3D%3D",
-                    SearchFilter.Channel => "EgIQAg%3D%3D",
-                    _ => null
-                },
-                continuation = continuationToken,
-                context = new
-                {
-                    client = new
+                    "query": "{{searchQuery}}",
+                    "params": "{{searchFilter switch
                     {
-                        clientName = "WEB",
-                        clientVersion = "2.20210408.08.00",
-                        hl = "en",
-                        gl = "US",
-                        utcOffsetMinutes = 0
+                        SearchFilter.Video => "EgIQAQ%3D%3D",
+                        SearchFilter.Playlist => "EgIQAw%3D%3D",
+                        SearchFilter.Channel => "EgIQAg%3D%3D",
+                        _ => null
+                    }}}",
+                    "continuation": "{{continuationToken}}",
+                    "context": {
+                        "client": {
+                            "clientName": "WEB",
+                            "clientVersion": "2.20210408.08.00",
+                            "hl": "en",
+                            "gl": "US",
+                            "utcOffsetMinutes": 0
+                        }
                     }
                 }
-            })
+                """
+            )
         };
 
         using var response = await _http.SendAsync(request, cancellationToken);