+++ YoutubeExplode/Videos/ClosedCaptions/ClosedCaptionClient.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.IO;
-using System.Linq;
 using System.Net.Http;
 using System.Runtime.CompilerServices;
 using System.Text;
@@ -28,7 +28,8 @@ public class ClosedCaptionClient
         VideoId videoId,
         [EnumeratorCancellation] CancellationToken cancellationToken = default)
     {
-        var playerResponse = await _controller.GetPlayerResponseAsync(videoId, cancellationToken);
+        // Use the TVHTML5 client instead of ANDROID_TESTSUITE because the latter doesn't provide closed captions
+        var playerResponse = await _controller.GetPlayerResponseAsync(videoId, null, cancellationToken);
 
         foreach (var trackData in playerResponse.ClosedCaptionTracks)
         {
@@ -68,10 +69,12 @@ public class ClosedCaptionClient
 
         foreach (var captionData in response.Captions)
         {
-            // Captions may have no text, but we should still include them to stay consistent
-            // with YouTube player behavior where captions are still displayed even if they're empty.
+            var text = captionData.Text;
+
+            // Skip over empty captions, but not captions containing only whitespace
             // https://github.com/Tyrrrz/YoutubeExplode/issues/671
-            var text = captionData.Text ?? "";
+            if (string.IsNullOrEmpty(text))
+                continue;
 
             // Auto-generated captions may be missing offset or duration.
             // https://github.com/Tyrrrz/YoutubeExplode/discussions/619
@@ -81,19 +84,24 @@ public class ClosedCaptionClient
                 continue;
             }
 
-            var parts = captionData.Parts.Select(p =>
+            var parts = new List<ClosedCaptionPart>();
+            foreach (var partData in captionData.Parts)
             {
-                // Caption parts may have no text, but we should still include them to stay consistent
-                // with YouTube player behavior where captions are still displayed even if they're empty.
+                var partText = partData.Text;
+
+                // Skip over empty parts, but not parts containing only whitespace
                 // https://github.com/Tyrrrz/YoutubeExplode/issues/671
-                var partText = p.Text ?? "";
+                if (string.IsNullOrEmpty(partText))
+                    continue;
 
                 var partOffset =
-                    p.Offset ??
+                    partData.Offset ??
                     throw new YoutubeExplodeException("Could not extract caption part offset.");
 
-                return new ClosedCaptionPart(partText, partOffset);
-            }).ToArray();
+                var part = new ClosedCaptionPart(partText, partOffset);
+
+                parts.Add(part);
+            }
 
             yield return new ClosedCaption(
                 text,
@@ -124,29 +132,35 @@ public class ClosedCaptionClient
         IProgress<double>? progress = null,
         CancellationToken cancellationToken = default)
     {
+        static string FormatTimestamp(TimeSpan value) =>
+            Math.Floor(value.TotalHours).ToString("00", CultureInfo.InvariantCulture) + ':' +
+            value.Minutes.ToString("00", CultureInfo.InvariantCulture) + ':' +
+            value.Seconds.ToString("00", CultureInfo.InvariantCulture) + ',' +
+            value.Milliseconds.ToString("000", CultureInfo.InvariantCulture);
+
+        // Would be better to use GetClosedCaptionsAsync(...) instead for streaming,
+        // but we need the total number of captions to report progress.
         var track = await GetAsync(trackInfo, cancellationToken);
 
         var buffer = new StringBuilder();
-        for (var i = 0; i < track.Captions.Count; i++)
+        foreach (var (caption, i) in track.Captions.WithIndex())
         {
-            var caption = track.Captions[i];
-            buffer.Clear();
-
             cancellationToken.ThrowIfCancellationRequested();
 
-            // Line number
-            buffer.AppendLine((i + 1).ToString());
-
-            // Time start --> time end
-            buffer.Append(caption.Offset.ToString(@"hh\:mm\:ss\,fff"));
-            buffer.Append(" --> ");
-            buffer.Append((caption.Offset + caption.Duration).ToString(@"hh\:mm\:ss\,fff"));
-            buffer.AppendLine();
-
-            // Actual text
-            buffer.AppendLine(caption.Text);
+            buffer
+                // Line number
+                .AppendLine((i + 1).ToString(CultureInfo.InvariantCulture))
+                // Time start --> time end
+                .Append(FormatTimestamp(caption.Offset))
+                .Append(" --> ")
+                .Append(FormatTimestamp(caption.Offset + caption.Duration))
+                .AppendLine()
+                // Content
+                .AppendLine(caption.Text);
 
             await writer.WriteLineAsync(buffer.ToString());
+            buffer.Clear();
+
             progress?.Report((i + 1.0) / track.Captions.Count);
         }
     }