+++ YoutubeExplode/Videos/VideoClient.cs
@@ -52,7 +52,9 @@ public class VideoClient
 
         var title =
             playerResponse.Title ??
-            throw new YoutubeExplodeException("Could not extract video title.");
+            // Videos without title are legal
+            // https://github.com/Tyrrrz/YoutubeExplode/issues/700
+            "";
 
         var channelTitle =
             playerResponse.Author ??
@@ -64,6 +66,7 @@ public class VideoClient
 
         var uploadDate =
             playerResponse.UploadDate ??
+            watchPage.UploadDate ??
             throw new YoutubeExplodeException("Could not extract video upload date.");
 
         var thumbnails = playerResponse.Thumbnails.Select(t =>